Converting ShieldBot to Sumo Robot League's MRK1
Mostly superficial changes converting the existing Shieldbot to MRK1 in
order to get quickkly to having a "proof of concept" to show people.

README.md was changed to reflect the connection of this project to Sumo
Robot League

In Index.html removed the links to www.makewitharduino.com and to the
Chrome App download, and others.

Took a SumoRobot logo and named it Shieldbot.jpg replacing the ShieldBot
image.